requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
shap>=0.41.0

# Deep learning
torch>=1.9.0
torchvision>=0.10.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Data processing
h5py>=3.1.0
tqdm>=4.62.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Experiment tracking
tensorboard>=2.7.0
wandb>=0.12.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Code quality
black>=21.0.0
flake8>=4.0.0
mypy>=0.910

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.4.0
ipywidgets>=7.6.0

# Utils
click>=8.0.0
pathlib2>=2.3.0
